Model: Fixed get method of propertyHeightInches to return int

+++ src/Model/Model.java
@@ -265,7 +265,7 @@ public class Model {
 		this.prefControl = prefControl;
 	}
 
-	public double getPropertyHeightInches() {
+	public int getPropertyHeightInches() {
 		return propertyHeightInches;
 	}
